requirements.txt
asgiref==3.7.2
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
